# Office-management-railway-deployment/swagger.yaml
openapi: 3.0.0
info:
  title: Auto Computation Office Management API
  version: 3.0.0
  description: API documentation for the Office Management System Backend
servers:
  - url: http://localhost:3000
    description: Local server
  - url: https://office-management-sbuh.onrender.com
    description: Production server (Render)
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token stored in a cookie
security:
  - cookieAuth: []
paths:
...
  /admin/emp/all:
    get:
      summary: Get all active employees
      tags:
        - Admin-Employees
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of active employees
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      type: object
        "500":
          description: Server error
...
  /admin/leaves/all:
    get:
      summary: Get all leave requests
      tags:
        - Admin-Leaves
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of leave requests
        "500":
          description: Server error
...
  /auth/login:
    post:
      summary: User login
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
                forceLogout:
                  type: boolean
      responses:
        "200":
          description: Login successful
        "401":
          description: Invalid credentials
        "500":
          description: Server error
  /auth/forgot-password:
    post:
      summary: Request a password reset
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
      responses:
        "200":
          description: Password reset email sent
        "404":
          description: User not found
        "500":
          description: Internal server error
  /auth/reset-password:
    post:
      summary: Reset password using token
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - password
              properties:
                token:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Password reset successfully
        "400":
          description: Invalid or expired token
        "500":
          description: Internal server error
  /auth/me:
    get:
      summary: Get currently logged in user info (from token)
      tags:
        - Auth
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Decoded user info
        "401":
          description: Not authenticated or invalid token
  /auth/myData:
    get:
      summary: Get current user full profile from DB
      tags:
        - Auth
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Full user profile
        "401":
          description: Not authenticated or invalid token
  /auth/logout:
    post:
      summary: Logout user and clear session
      tags:
        - Auth
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Logout successful
        "500":
          description: Server error
  /auth/2fa:
    get:
      summary: Check if 2FA is enabled for the current user
      tags:
        - Auth
      security:
        - cookieAuth: []
      responses:
        "200":
          description: 2FA status
  /auth/authorized-2fa:
    post:
      summary: Verify 2FA code during login
      tags:
        - Auth
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
      responses:
        "200":
          description: 2FA verification successful
        "401":
          description: Invalid 2FA code or unauthorized
        "500":
          description: Server error
  /auth/verify-password:
    post:
      summary: Verify current user password
      tags:
        - Auth
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
      responses:
        "200":
          description: Password verified
        "401":
          description: Invalid password or unauthorized
        "500":
          description: Server error
...
  /employee/leaves/apply:
    post:
      summary: Apply for a new leave
      tags:
        - Employee-Leaves
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - start_date
                - end_date
                - reason
              properties:
                type:
                  type: string
                  enum:
                    - Sick
                    - Casual
                    - Privilege
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
                reason:
                  type: string
      responses:
        "201":
          description: Leave application submitted successfully
        "400":
          description: Missing required fields or invalid leave type
        "409":
          description: Leave request already exists for these dates
        "500":
          description: Server error
  /employee/leaves/summary:
    get:
      summary: Get leave entitlement summary and balance
      tags:
        - Employee-Leaves
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of leave types with used and available balances
        "500":
          description: Server error
  /employee/leaves/all:
    get:
      summary: Get all leave applications for the current employee
      tags:
        - Employee-Leaves
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of all leaves
        "500":
          description: Server error
...